devapi/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
oracledb==2.3.0
python-dotenv==1.0.1
pydantic[email]==2.7.1
